backend/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
pinecone
langchain
sentence-transformers
pypdf
google-generativeai
